Rename `TermsOfUseCard` file name
