+++ src/pages/components/OrdinoHelpSection.tsx
@@ -5,7 +5,7 @@ import {IVideoCardProps, VideoCard} from './VideoCard';
 import {OrdinoScrollspy, OrdinoScrollspyItem, OrdinoFooter} from 'ordino';
 import {UniqueIdManager} from 'phovea_core';
 import {SourceCodeCard} from './SourceCodeCard';
-import {TermsOfUseCard} from './TermsOfuseCard';
+import {TermsOfUseCard} from './TermsOfUseCard';
 
 const cards = [
   {
