+++ src/pages/components/TermsOfUseCard.tsx
